refactor(device): centralize physical memory probing
Move physical-memory discovery into one shared implementation.
Give setup and inference code a consistent RAM capacity value.
